+++ typings/index.d.ts
@@ -1,17 +1,3 @@
 export { ShewenyClient } from "./ShewenyClient";
-export {
-  MessageCommandsHandler,
-  ButtonsHandler,
-  EventsHandler,
-  SelectMenusHandler,
-  ApplicationCommandsHandler,
-  InhibitorsHandler,
-} from "./handlers";
-export {
-  MessageCommand,
-  Button,
-  Event,
-  SelectMenu,
-  Inhibitor,
-  ApplicationCommand,
-} from "./structures";
+export { MessageCommandsHandler, ButtonsHandler, EventsHandler, SelectMenusHandler, ApplicationCommandsHandler, InhibitorsHandler, } from "./handlers";
+export { MessageCommand, Button, Event, SelectMenu, Inhibitor, ApplicationCommand, } from "./structures";
