+++ typings/typescript/interfaces/interfaces.d.ts
@@ -2,3 +2,7 @@ export interface IMessageCommandHandlerOptions {
     directory: string;
     prefix?: string;
 }
+export interface ILoadAllApplicationCommand {
+    loadAll: boolean;
+    guildId?: string;
+}
