--- typings/typescript/types.d.ts
@@ -1,4 +1,0 @@
-/**
- * Inhibitors types
- */
-export declare type InhibitorType = "COMMAND" & "BUTTON" & "SELECT_MENU";
